pyproject.toml
[project]
name = "airflow-operator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-client==2.10.0",
    "kopf>=1.38.0",
    "google-auth[requests]>=2.43.0",
    "kubernetes>=34.1.0",
    "opentelemetry-instrumentation-requests==0.59b0",
    "prometheus-client==0.23.1",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
]
